+++ vkconfig_core/command_line.cpp
@@ -37,10 +37,7 @@ struct CommandHelpDesc {
 };
 
 static const CommandHelpDesc command_help_desc[] = {
-    {HELP_HELP, "help"},
-    {HELP_VERSION, "version"},
-    {HELP_GUI, "gui"},
-    {HELP_LAYERS, "layers"},
+    {HELP_HELP, "help"}, {HELP_VERSION, "version"}, {HELP_GUI, "gui"}, {HELP_LAYERS, "layers"}, {HELP_RESET, "reset"},
 };
 
 static HelpType GetCommandHelpId(const char* token) {
@@ -69,6 +66,7 @@ static const ModeDesc mode_desc[] = {
     {COMMAND_VERSION, "version", HELP_VERSION},    // COMMAND_VERSION
     {COMMAND_GUI, "gui", HELP_GUI},                // COMMAND_GUI
     {COMMAND_LAYERS, "layers", HELP_LAYERS},       // COMMAND_LAYERS
+    {COMMAND_RESET, "reset", HELP_RESET}           // COMMAND_RESET
 };
 
 static CommandType GetModeId(const char* token) {
@@ -88,6 +86,38 @@ static const ModeDesc& GetModeDesc(CommandType command_type) {
 
     assert(0);
     return mode_desc[0];  // COMMAND_NONE
+}
+
+struct CommandResetDesc {
+    CommandResetArg arguments;
+    const char* token;
+    int required_arguments;
+};
+
+static const CommandResetDesc command_reset_desc[] = {{COMMAND_RESET_SOFT, "--soft", 2},
+                                                      {COMMAND_RESET_SOFT, "-s", 2},
+                                                      {COMMAND_RESET_HARD, "--hard", 2},
+                                                      {COMMAND_RESET_HARD, "-h", 2}};
+
+static CommandResetArg GetCommandResetId(const char* token) {
+    assert(token != nullptr);
+
+    for (std::size_t i = 0, n = countof(command_reset_desc); i < n; ++i) {
+        if (std::strcmp(command_reset_desc[i].token, token) == 0) return command_reset_desc[i].arguments;
+    }
+
+    return COMMAND_RESET_NONE;
+}
+
+static const CommandResetDesc& GetCommandReset(CommandResetArg reset_arg) {
+    assert(reset_arg != COMMAND_RESET_NONE);
+
+    for (std::size_t i = 0, n = countof(command_reset_desc); i < n; ++i) {
+        if (command_reset_desc[i].arguments == reset_arg) return command_reset_desc[i];
+    }
+
+    assert(0);
+    return command_reset_desc[0];
 }
 
 struct CommandLayersDesc {
@@ -130,11 +160,13 @@ static const CommandLayersDesc& GetCommandLayers(CommandLayersArg layers_arg) {
 
 CommandLine::CommandLine(int argc, char* argv[])
     : command(_command),
+      command_reset_arg(_command_reset_arg),
       command_layers_arg(_command_layers_arg),
       layers_configuration_path(_layers_configuration_path),
       error(_error),
       error_args(_error_args),
       _command(COMMAND_GUI),
+      _command_reset_arg(COMMAND_RESET_NONE),
       _command_layers_arg(COMMAND_LAYERS_NONE),
       _error(ERROR_NONE),
       _help(HELP_DEFAULT) {
@@ -178,6 +210,20 @@ CommandLine::CommandLine(int argc, char* argv[])
                     _error = ERROR_FILE_NOTFOUND;
                     _error_args.push_back(argv[arg_offset + 2]);
                 }
+                break;
+            }
+        } break;
+        case COMMAND_RESET: {
+            if (argc <= arg_offset + 1) {
+                _command_reset_arg = COMMAND_RESET_SOFT;
+                break;
+            }
+
+            _command_reset_arg = GetCommandResetId(argv[arg_offset + 1]);
+            if (_command_reset_arg == COMMAND_RESET_NONE) {
+                _error = ERROR_INVALID_COMMAND_ARGUMENT;
+                _error_args.push_back(argv[arg_offset + 0]);
+                _error_args.push_back(argv[arg_offset + 1]);
                 break;
             }
         } break;
@@ -262,6 +308,7 @@ void CommandLine::usage() const {
             printf("\tversion                   = Display %s version.\n", VKCONFIG_NAME);
             printf("\tgui                       = Launch the %s GUI.\n", VKCONFIG_NAME);
             printf("\tlayers                    = Manage system Vulkan Layers.\n");
+            printf("\treset                     = Reset layers configurations.\n");
             printf("\n");
             printf("  (Use 'vkconfig help <commamd>' for detailed usage of %s commands.)\n", VKCONFIG_NAME);
             break;
@@ -312,6 +359,23 @@ void CommandLine::usage() const {
             printf("\n");
             printf("\tvkconfig layers (--list-version | -lv)\n");
             printf("\t\tList the Vulkan layers found by %s on the system with locations and versions.\n", VKCONFIG_NAME);
+            break;
+        }
+        case HELP_RESET: {
+            printf("Name\n");
+            printf("\t'reset' - Reset layers configurations\n");
+            printf("\n");
+            printf("Synopsis\n");
+            printf("\tvkconfig reset (--hard | -h)\n");
+            printf("\tvkconfig reset <--soft | -s>\n");
+            printf("\n");
+            printf("Description\n");
+            printf("\tvkconfig reset (--hard | -h)\n");
+            printf("\t\tReset all layers configurations, all user-defined configurations will be lost.\n");
+            printf("\n");
+            printf("\tvkconfig reset <--soft | -s>\n");
+            printf("\t\tReset all default configurations, all user-defined configuration will be preserved.\n");
+            printf("\n");
             break;
         }
     }